requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
torch==2.1.1
monai==1.3.0
segment-anything==1.0
scipy==1.11.3
requests==2.31.0
matplotlib==3.8.2
python-dotenv==1.0.0
tqdm==4.66.1
